Fix logicalerr3.c: cast to double for floating-point division

+++ logicalerr3.c
@@ -2,7 +2,7 @@
 
 int main(void) {
     int a = 5, b = 2;
-    double result = a / b;   // logical error
+    double result = (double)a / b;   // fixed: cast to double for floating-point division
 
     printf("Result = %.2f\n", result);
     return 0;
